fix: do not search PATH for command on mgmt node
Previously dsh would use search_path() to find the fullpath
of the command, and then use fullpath as remote command.
This method fails when the management node runs a different OS
than the other node(s), where the command may be installed
in a different path. Notable example is differences between
distros that use /usr/bin vs /bin. Heck, PATH may be set to /Apps
or whatever. So, just pass the command and let it use the
target node's root user's PATH.